+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - Update proposer to leader and add leader name avatar in latest blocks (home) ([\#66](https://github.com/forbole/big-dipper-2.0-solana/issues/66))
 - Add: show number of validators in tabs title (validators page) ([\#68](https://github.com/forbole/big-dipper-2.0-solana/issues/68))
 - Add tokens chart ([\#76](https://github.com/forbole/big-dipper-2.0-solana/issues/76))
+- Add Stake Weight Component ([\#41](https://github.com/forbole/big-dipper-2.0-solana/issues/41))